Eventstreams with KafkaSSE initial commit on service-template
Routes are defined in spec.yaml, but must be also configured
in config.yaml.  This duplicates things a bit, but
in production we can render both a spec.yaml and a config.yaml
from the same single source.

Note: It is difficult to test streams, since they are HTTP requests
with no end, and spec tests use preq by default.  So, we only
test the stream endpoints that should return 404s.  Stream
testing is left up to the kafka-sse module.

Bug: T148779
Change-Id: If2b57914cc88953e7dc98748e85c51df87d22848