Restrict topics in a stream to those configured for that stream
This restricts the ability of users to set Last-Event-ID to consume
topics from a route that are not configured for that stream route.

This doesn't actually increase security at all.  It just enforces
stream route semantics, to keep people from accidentally consuming
other (allowed) topics from a certain stream route.

Change-Id: Ibe4260c65e5736156e194bf362062c98dc67fb9f