Allow to configure EventStreams to consume any topic
This will allow us to set up an EventStreams instance that
can consume from any topic in Kafka.  This should not
be enabled in a production environment!

This will be used by WMF to set up an instance of EventStreams
in our beta environment that will allow for easy testing
and debugging of dynamic streams.

Bug: T253069
Change-Id: I541bfa2ecb92a6910bfb709ae6a4ec46e3ce4b21