stream.js: add extra logging for client connect/disconnect
We don't have a lot of visibility on what HTTP metadata
the HTTP clients of ES are carrying, more specifically
User Agent and Referer. This change adds basic logging
when a new client connects and disconnects, to be able
to better follow up when certain changes to streams
are required (say deprecating one).

We cannot use Varnish Webrequest data since we configured
it to act as straight TCP proxy.

Remove also the client-ip label from Prometheus metrics.

Bug: T343511
Change-Id: I6c08652265d76a786b8662b4571789d8607aa469